feat: respect deepLevel, order keys, new property showDownloadButton, refactor download code
